requirements.txt
absl-py==0.5.0
astor==0.7.1
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
gast==0.2.0
Glances==3.0.2
grpcio==1.15.0
h5py==2.8.0
idna==2.7
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.2
pandas==0.23.4
Pillow==5.3.0
pkginfo==1.4.2
protobuf==3.6.1
psutil==5.4.7
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
readme-renderer==24.0
requests==2.20.0
requests-toolbelt==0.8.0
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-hub==0.1.1
termcolor==1.1.0
tfcoreml==0.3.0
tqdm==4.28.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
